requirements.txt
Django==1.9.1
django-push-notifications==1.4.1
selenium==2.50.1
wheel==0.24.0
